Move isLayoutToolHidden from reducer to selector
