+++ test/unit/reducers/gui_test.js
@@ -29,7 +29,6 @@ describe('gui reducer', () => {
       'isAuthenticationView',
       'isCompleterActive',
       'isGridMode',
-      'isLayoutToolHidden',
       'isNavbarHidden',
       'isNotificationsActive',
       'isNotificationsUnread',
@@ -240,19 +239,6 @@ describe('gui reducer', () => {
 
       const action2 = { type: LOCATION_CHANGE, payload: { pathname: '/starred' } }
       expect(reducer(undefined, action2)).to.have.property('isGridMode', false)
-    })
-
-    it('LOCATION_CHANGE updates whether the layout tool is displayed', () => {
-      const action1 = { type: LOCATION_CHANGE, payload: { pathname: '/following' } }
-      expect(reducer(undefined, action1)).to.have.property('isLayoutToolHidden', false)
-      const action2 = { type: LOCATION_CHANGE, payload: { pathname: '/settings' } }
-      expect(reducer(undefined, action2)).to.have.property('isLayoutToolHidden', true)
-      const action3 = { type: LOCATION_CHANGE, payload: { pathname: '/elloindiegogo' } }
-      expect(reducer(undefined, action3)).to.have.property('isLayoutToolHidden', false)
-      const action4 = { type: LOCATION_CHANGE, payload: { pathname: '/elloindiegogo/post/a5qr9a' } }
-      expect(reducer(undefined, action4)).to.have.property('isLayoutToolHidden', true)
-      const action5 = { type: LOCATION_CHANGE, payload: { pathname: '/discover/all' } }
-      expect(reducer(undefined, action5)).to.have.property('isLayoutToolHidden', true)
     })
 
     it('LOCATION_CHANGE updates isAuthenticationView', () => {
